Fix #18: extra json file causes blank pages at the end

+++ pdf2rmnotebook.sh
@@ -18,11 +18,38 @@
 #                1.2.0 : Usage, Verbosity and Version options
 #                1.3.0 : Image inclusion
 #                2.2.0 : option to create reMarkable Notebook .rmn file
+#                3.0.0 : Color management for Pro, rmdoc
+#                3.1.0 : Color management for text files imorted
 #===============================================================================
+
+#  reMarkable 2: device screen width ≅ 157 mm, height = 209 mm
+#  reMarkable 2: monochrome display, drawj2d will map colours to black, grey or white
+#  reMarkable Pro: device screen width ≅ 179 mm, height = 239 mm.
+#     Preview drawj2d -Tscreen -W168 -H239 thales.hcl
+#
+#  reMarkable 2:
+# Embed page 5 of an A4 sized (297mm x 210mm) pdf ﬁle, scale to ﬁt the tablet height (297 * 0.7 = 208mm < 209mm), right justiﬁed (9 + 210 * 0.7 = 156mm < 157mm).
+#    moveto 9 0
+#    image article.pdf 5 0 0 0.7
+#
+# reMarkable Pro:
+# Embed page 5 of an A4 sized (297mm x 210mm) pdf ﬁle, scale to ﬁt the tablet height (297 * 0.8 = 238mm < 239mm), right justiﬁed (210*0.8 = 168mm <= 179mm - 11mm).
+#     move10 11 0
+#     image article.pdf 5 0 0 0.8
+#  convert pappagallo.jpg pappagallo.pdf
+#  convert pappagallo.pdf -dither FloydSteinberg -remap colormap.png rM-pappagallo.pdf
+
+# From: https://github.com/ricklupton/rmc
+#    Command line tool for converting to/from remarkable .rm version 6 (software version 3) files.
+# ~/Python-Env/rmc/bin/rmc -t rm  -o GrafanaMail.rm GrafanaMail.md
+
+# To add multiple files in multiple colors arg parsing should be splitted,
+# everything will become too complicated. please use multiple sepaate
+# conversions, merge on the device.
 
 set -o nounset                              # Treat unset variables as an error
 
-Version=2.2.0
+Version=3.1.0
 
 NAME=$(basename $0 .sh)
 TEMP=$(mktemp -d)
@@ -37,20 +64,22 @@ Usage:
 Create multi-page reMarkable Notebook file from PDF files
   * Creates .zip files by default for use with rmapi
   * Use -r option to create a reMarkable Notebook .rmn file for use with RCU
+  * Use -R option to create a reMarkable Notebook .rmdoc file for use with USB Link
 
-Options:
   Switches (no arguments required):
     -h    Display this help and exit
     -q    Produce fewer messages to stdout
     -r    Create a reMarkable Notebook .rmn file (default: zip)
+    -R    Create a reMarkable Notebook .rmdoc file
     -v    Produce more messages to stdout
     -V    Display version information and exit
 
   With arguments:
     -n NAME    Set the rmn Notebook Display Name (default: Notebook-<yyyymmdd_hhmm.ss>)
-               Only used with -r option
+               Only used with -r/-R option
     -o FILE    Set the output filename (default: Notebook-<yyyymmdd_hhmm.ss>.zip)
     -s SCALE   Set the scale value (default: 0.75) - 0.75 is a good value for A4/Letter PDFs
+    -c COLOR   Set a color for the files to be imported
 
 Example:
   $NAME -n "My Notebook" -o mynotebook.zip -s 1.0 file.pdf
@@ -59,6 +88,30 @@ EOD
 
 Cleanup() {
  rm -rf ${TEMP}
+}
+
+# Simple ASCII text file is embedded into HCL and then converted
+textFile(){
+  # $1 = Filename
+  # $2 = HCL Page
+
+  _F=$1
+  _HCL=$2
+
+  # Overwrite default page settings
+    echo pen $COLOR 0.1 solid > ${_HCL}
+    echo font Lines up 3.5 >> ${_HCL}
+    echo moveto 12 8 >> ${_HCL}
+
+    cat ${_F} >> ${_HCL}.tmp
+    sed -i 's/"/\\"/g'  ${_HCL}.tmp
+    sed -i 's/`/\\`/g'  ${_HCL}.tmp
+    sed -i 's/]/\\]/g' ${_HCL}.tmp
+    sed -i 's/\[/\\[/g' ${_HCL}.tmp
+    sed -i 's/\$/\\$/g' ${_HCL}.tmp
+    sed -i 's/\(.*\)/text "\1" 140/' ${_HCL}.tmp
+    cat ${_HCL}.tmp >> ${_HCL}
+    rm ${_HCL}.tmp
 }
 
 # === Main ===
@@ -74,10 +127,21 @@ OUTFILE=""
 EXTENSION=".zip"
 DISPLAY_NAME=$DEFAULT_OUTFILE
 RMN=false
+RMDOC=false
+RMC=false
+COLOR=black
 
-while getopts "dhin:qrvs:Vo:" opt
+while getopts "c:dhimn:qrRvs:Vo:" opt
 do
   case $opt in
+    c)
+      COLOR=$OPTARG
+      ;;
+    m)
+      # RMC Converted file
+      RMC=true
+      EXTENSION=".rmdoc"
+      ;;
     d)
       # Fake some values
       DEBUG=true
@@ -97,6 +161,10 @@ do
       RMN=true
       EXTENSION=".rmn"
       TARARGS="cf"
+      ;;
+    R)
+      RMDOC=true
+      EXTENSION=".rmdoc"
       ;;
     h)
       Usage
@@ -147,7 +215,7 @@ fi
 
 # Prepare HCL file for image inclusion
 cat > ${TEMP}/page.hcl <<EOF
-pen black 0.1 solid
+pen $COLOR 0.1 solid
 line 1 1 156 1
 line 156 1 156 208
 line 156 208 1 208
@@ -162,7 +230,7 @@ _page=0
 UUID_N=$(uuidgen)   # UUID for Notebook
 
 mkdir ${NB}/${UUID_N}
-cp ${VARLIB}/UUID.pagedata ${NB}/${UUID_N}.pagedata
+#cp ${VARLIB}/UUID.pagedata ${NB}/${UUID_N}.pagedata
 cp ${VARLIB}/UUID_HEAD.content ${NB}/${UUID_N}.content
 
 for _P in "$@"
@@ -174,6 +242,17 @@ do
   test -f "${_P}" || { echo "${_P}: No such file or directory." ; Usage ; exit 1 ; }
 
   FILETYPE=$(file -b --mime-type "${_P}")
+  if [ -z "$FILETYPE" -o "$FILETYPE" = "text/plain" ]
+  then
+    EXT=${_P##*.}
+    case $EXT in
+      md)   # Markdown
+        RMC=true
+        FILETYPE='text/markdown'
+        _NP=1
+        ;;
+    esac
+  fi
 
   case $FILETYPE in
     image/jpeg | image/png)
@@ -184,7 +263,7 @@ do
     application/pdf )
 
       # Get Pages from file, loop over all of them
-      read x _NP <<< $(pdfinfo "${_P}" | grep Pages: )
+      read x _NP <<< $(pdfinfo "${_P}" | grep -a Pages: )
       ;;
   esac
 
@@ -203,6 +282,9 @@ do
           # Scale image: Usually PDF is A4/Letter, rM is smaller
           echo "image {${_P}} ${_PP} 0 0 $SCALE" >> ${TEMP}/P_${_page}.hcl
           ;;
+        text/markdown )
+          textFile ${_P} ${TEMP}/P_${_page}.hcl
+          ;;
       esac
 
       UUID_P=$(uuidgen)   # Notebook Pages should be named using the UUID from .content file
@@ -210,9 +292,17 @@ do
   #  so we need to use the same convention in naming pages: 0.rm 1.rm ... instead of UUIDs
   #  which are used internally in the rM (see ~/.local/share/remarkable/xochitl/ )
   #  It is indeed easier to have page numbers instead of random UUIDs referenced elsewhere
-      drawj2d ${QVFLAG} -T rm -o ${NB}/${UUID_N}/${_page}.rm ${TEMP}/P_${_page}.hcl
-# Fix issue #18
-#     cp ${VARLIB}/UUID_PAGE-metadata.json ${NB}/${UUID_N}/${_page}-metadata.json
+  # Fix issue #18: remove .json file
+      if [ ! $RMDOC ]
+      then
+        drawj2d ${QVFLAG} -T rm -o ${NB}/${UUID_N}/${_page}.rm ${TEMP}/P_${_page}.hcl
+#        cp ${VARLIB}/UUID_PAGE-metadata.json ${NB}/${UUID_N}/${_page}-metadata.json
+      else
+$DEBUG && cat ${TEMP}/P_${_page}.hcl
+$DEBUG && echo drawj2d ${QVFLAG} -T rm -o ${NB}/${UUID_N}/${UUID_P}.rm ${TEMP}/P_${_page}.hcl
+        drawj2d ${QVFLAG} -T rm -o ${NB}/${UUID_N}/${UUID_P}.rm ${TEMP}/P_${_page}.hcl
+#        cp ${VARLIB}/UUID_PAGE-metadata.json ${NB}/${UUID_N}/${UUID_P}-metadata.json
+      fi
 
       echo -n \"${UUID_P}\" >> ${NB}/${UUID_N}.content
 
@@ -234,7 +324,6 @@ fi
 
 (
 cd ${NB}
-if [ $RMN = true ]; then
   if [ $QVFLAG = "-v" ]; then
     TARARGS="${TARARGS}v"
   fi
@@ -260,6 +349,7 @@ if [ $RMN = true ]; then
   }
 EOF
 
+if [ $RMN = true ]; then
   tar $TARARGS ${TEMP}/Notebook$EXTENSION ${UUID_N}.* ${UUID_N}/*
 else
   zip ${QVFLAG} ${TEMP}/Notebook$EXTENSION ${UUID_N}.* ${UUID_N}/*
@@ -272,4 +362,6 @@ cp ${TEMP}/Notebook$EXTENSION ${OUTFILE}$EXTENSION
 echo Output written to $OUTFILE$EXTENSION
 
 #DEBUG  find ${TEMP} -ls
+
+# vim:set ai et sts=2 sw=2 tw=80:
